Organize files and fix bundle id issue
