+++ plugins/notification-service-extension/plugin/src/iosConstants.ts
@@ -17,7 +17,8 @@ export const NSE_SWIFT_SOURCE_FILES = [
   "XmtpClient.swift",
   "PushNotificationContentFactory.swift",
   "XMTPContentDecoder.swift",
-  "Group+MemberNames.swift"
+  "Group+MemberNames.swift",
+  "Bundle+Helpers.swift"
 ]
 export const NSE_INFO_PLIST_FILE_NAME = `${NSE_TARGET_NAME}-Info.plist`
 export const NSE_EXT_FILES = [`${NSE_TARGET_NAME}.entitlements`, NSE_INFO_PLIST_FILE_NAME]
